compute minimal adjustment sets and pick one deterministically

compute minimal adjustment sets and pick one deterministically

Usage

pick_minimal_controls(dag, exposure, outcome)

Arguments

dag

A dagitty DAG object.

exposure

Character; exposure node name (X).

outcome

Character; outcome node name (Y).

Value

sorted character vector with the minimal adjustment set

Details

This is preferable to dagitty::adjustmentSets() because it picks a single minimal adjustment set automatically and without a bunch of arguments. will need to adapt this to deal with multiple adjustment sets and canonical adjustment sets eventually.

Examples

d <- ggdag::dagify(
Y ~ X + M + Z,
M ~ X + Z,
C ~ X + Y,
exposure = "X",
outcome = "Y")
# M: mediator / Z: confounder / C: collider

# identify the minimal adjustment set
pick_minimal_controls(d, exposure = "X", outcome = "Y")
#> character(0)
# expected: "Z"